Move parse_lines method to loadlog
It's only used for that endpoint
