+++ bluebird/api/resources/utils.py
@@ -176,56 +176,3 @@ def validate_scenario(scn_lines):
 			return f'Line \'{line}\' does not match the required format'
 
 	return None
-
-
-def parse_lines(lines, target_time=0):
-	"""
-	Parses the content of an episode file
-	:param lines:
-	:param target_time:
-	:return: String for errors or dict with parsed data
-	"""
-
-	if not lines:
-		return 'No lines provided'
-
-	lines = list(lines)  # Take a copy of the input
-
-	# Get the seed from the first line
-	match = re.match(r'.*Episode started.*Seed is (\d+)', lines.pop(0))
-	if not match:
-		return 'Episode seed was not set'
-	scn_data = {'seed': int(match.group(1)), 'lines': []}
-
-	if not lines:
-		return 'No more lines after parsing seed'
-
-	while not 'Scenario file loaded' in lines[0]:
-		lines.pop(0)
-		if not lines:
-			return 'Couldn\'t find scenario content'
-
-	lines.pop(0)
-
-	while lines:
-		match = re.match(r'.*E.*>.*', lines[0])
-		if not match:
-			break
-		scn_data['lines'].append(lines.pop(0).split(' E ')[1])
-
-	if not lines:
-		return 'No more lines after parsing scenario content'
-
-	while lines:
-		line = lines.pop(0)
-		if 'Episode finished' in line:
-			return scn_data
-		match = re.match(r'.*C \[(\d+)\] (.*)$', line)
-		if match and not 'STEP' in line:
-			time_s = int(match.group(1))
-			if not target_time or time_s <= target_time:
-				cmd_time = time.strftime('%H:%M:%S', time.gmtime(time_s))
-				cmd_str = match.group(2)
-				scn_data['lines'].append(f'{cmd_time}> {cmd_str}')
-
-	return 'Could not find end of episode'
